2141: Requested changes about refactoring frinet data syncing
