+++ administration/src/graphql/freinet/sendApplicationAndCardDataToFreinet.graphql
@@ -1,6 +1,3 @@
 mutation sendApplicationDataToFreinet($applicationId: Int!, $project: String!) {
-  result: sendApplicationDataToFreinet(applicationId: $applicationId, project: $project) {
-    success
-    errorMessage
-  }
+  sendApplicationDataToFreinet(applicationId: $applicationId, project: $project)
 }
