+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/exception/webservice/exceptions/FreinetApplicationSyncFailedException.kt
@@ -1,0 +1,8 @@
+package app.ehrenamtskarte.backend.exception.webservice.exceptions
+
+import app.ehrenamtskarte.backend.exception.GraphQLBaseException
+import app.ehrenamtskarte.backend.exception.webservice.schema.GraphQLExceptionCode
+
+class FreinetApplicationSyncFailedException : GraphQLBaseException(
+    GraphQLExceptionCode.FREINET_APPLICATION_SYNC_FAILED,
+)
