+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/exception/webservice/exceptions/FreinetDataTransfernApiNotReachableException.kt
@@ -1,0 +1,8 @@
+package app.ehrenamtskarte.backend.exception.webservice.exceptions
+
+import app.ehrenamtskarte.backend.exception.GraphQLBaseException
+import app.ehrenamtskarte.backend.exception.webservice.schema.GraphQLExceptionCode
+
+class FreinetDataTransfernApiNotReachableException : GraphQLBaseException(
+    GraphQLExceptionCode.FREINET_DATA_TRANSFER_API_NOT_REACHABLE,
+)
