+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/exception/webservice/exceptions/FreinetFoundMultiplePersonsException.kt
@@ -1,0 +1,8 @@
+package app.ehrenamtskarte.backend.exception.webservice.exceptions
+
+import app.ehrenamtskarte.backend.exception.GraphQLBaseException
+import app.ehrenamtskarte.backend.exception.webservice.schema.GraphQLExceptionCode
+
+class FreinetFoundMultiplePersonsException : GraphQLBaseException(
+    GraphQLExceptionCode.FREINET_FOUND_MULTIPLE_PERSONS,
+)
