+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/exception/webservice/schema/GraphQLExceptionCode.kt
@@ -3,6 +3,10 @@ package app.ehrenamtskarte.backend.exception.webservice.schema
 enum class GraphQLExceptionCode {
     EMAIL_ALREADY_EXISTS,
     FREINET_AGENCY_NOT_FOUND,
+    FREINET_APPLICATION_SYNC_FAILED,
+    FREINET_DATA_TRANSFER_NOT_ACTIVATED,
+    FREINET_DATA_TRANSFER_API_NOT_REACHABLE,
+    FREINET_FOUND_MULTIPLE_PERSONS,
     INVALID_INPUT,
     INVALID_APPLICATION_CONFIRMATION_NOTE_SIZE,
     INVALID_CARD_HASH,
