--- backend/src/main/kotlin/app/ehrenamtskarte/backend/freinet/webservice/schema/types/SendToFreinetResponse.kt
@@ -1,6 +1,0 @@
-package app.ehrenamtskarte.backend.freinet.webservice.schema.types
-
-data class SendToFreinetResponse(
-    val success: Boolean,
-    val errorMessage: String? = null,
-)
